refactor hosting (#278)
