+++ sample/SkyApm.Sample.Frontend/Program.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Hosting;
+﻿using Microsoft.AspNetCore;
+using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Hosting;
 
 namespace SkyApm.Sample.Frontend
@@ -10,14 +11,21 @@ namespace SkyApm.Sample.Frontend
             BuildHost(args).Run();
         }
 
+#if NETCOREAPP2_1
+
+        public static IWebHost BuildHost(string[] args) =>
+            WebHost.CreateDefaultBuilder<Startup>(args)
+                .Build();
+
+#else
+
         public static IHost BuildHost(string[] args) =>
             Host.CreateDefaultBuilder(args)
                 .ConfigureWebHostDefaults(builder =>
                 {
-                    builder
-                        .UseStartup<Startup>()
-                        .UseUrls("http://*:5001");
-                })
-                .Build();
+                    builder.UseStartup<Startup>();
+                }).Build();
+
+#endif
     }
 }
