+++ sample/SkyApm.Sample.Frontend/Startup.cs
@@ -4,6 +4,12 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using SkyApm.Sample.Backend.Services;
 
+#if NETCOREAPP2_1
+
+using IHostEnvironment = Microsoft.Extensions.Hosting.IHostingEnvironment;
+
+#endif
+
 namespace SkyApm.Sample.Frontend
 {
     public class Startup
@@ -18,7 +24,13 @@ namespace SkyApm.Sample.Frontend
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddControllers();
+#if NETCOREAPP2_1
+
+            services.AddMvc();
+
+#else
+             services.AddControllers();
+#endif
 
             // DI grpc service
             services.AddSingleton<GreeterGrpcService>();
@@ -31,13 +43,16 @@ namespace SkyApm.Sample.Frontend
             {
                 app.UseDeveloperExceptionPage();
             }
-
+#if NETCOREAPP2_1
+            app.UseMvcWithDefaultRoute();
+#else
             app.UseRouting();
 
             app.UseEndpoints(endpoint =>
             {
                 endpoint.MapDefaultControllerRoute();
             });
+#endif
         }
     }
 }
