+++ skyapm-dotnet.sln
@@ -84,10 +84,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkyApm.DotNet.CLI", "src\Sk
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkyApm.Diagnostics.SmartSql", "src\SkyApm.Diagnostics.SmartSql\SkyApm.Diagnostics.SmartSql.csproj", "{B43160C8-C9E3-43F6-83E9-A9A45B32F022}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkyApm.Agent.GeneralHost", "src\SkyApm.Agent.GeneralHost\SkyApm.Agent.GeneralHost.csproj", "{4C1CF0E7-4CD3-4731-9A56-E033D5216D58}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkyApm.Sample.GeneralHost", "sample\SkyApm.Sample.GeneralHost\SkyApm.Sample.GeneralHost.csproj", "{477D705E-576B-46C8-8F1E-9A86EDAE9D86}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkyApm.Benchmark", "benchmark\SkyApm.Benchmark\SkyApm.Benchmark.csproj", "{33581FDE-ABAF-4C27-A40A-1A2743309399}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "benchmark", "benchmark", "{2B7F59E8-147F-4399-9804-E7EAEF2DCB22}"
@@ -102,7 +98,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkyApm.Sample.GrpcProto", "
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkyApm.Sample.GrpcServer", "sample\grpc\SkyApm.Sample.GrpcServer\SkyApm.Sample.GrpcServer.csproj", "{0EB6D474-A915-4075-B142-FE07321F3AA9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SkyApm.Diagnostics.Grpc.Net.Client", "src\SkyApm.Diagnostics.Grpc.Net.Client\SkyApm.Diagnostics.Grpc.Net.Client.csproj", "{34211B07-381E-4489-AA7F-7697D106B7DC}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkyApm.Diagnostics.Grpc.Net.Client", "src\SkyApm.Diagnostics.Grpc.Net.Client\SkyApm.Diagnostics.Grpc.Net.Client.csproj", "{34211B07-381E-4489-AA7F-7697D106B7DC}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkyApm.Agent.Hosting", "src\SkyApm.Agent.Hosting\SkyApm.Agent.Hosting.csproj", "{BC56D244-6C76-4E40-8351-B74B636B17A6}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkyApm.Agent.GeneralHost", "src\SkyApm.Agent.GeneralHost\SkyApm.Agent.GeneralHost.csproj", "{D96279FD-0020-427B-85C5-DC1C2336AA25}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SkyApm.Sample.GenericHost", "sample\SkyApm.Sample.GenericHost\SkyApm.Sample.GenericHost.csproj", "{71F61C77-E04F-464D-A0B7-E109CE7E30D7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -198,14 +200,6 @@ Global
 		{B43160C8-C9E3-43F6-83E9-A9A45B32F022}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B43160C8-C9E3-43F6-83E9-A9A45B32F022}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B43160C8-C9E3-43F6-83E9-A9A45B32F022}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4C1CF0E7-4CD3-4731-9A56-E033D5216D58}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4C1CF0E7-4CD3-4731-9A56-E033D5216D58}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4C1CF0E7-4CD3-4731-9A56-E033D5216D58}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4C1CF0E7-4CD3-4731-9A56-E033D5216D58}.Release|Any CPU.Build.0 = Release|Any CPU
-		{477D705E-576B-46C8-8F1E-9A86EDAE9D86}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{477D705E-576B-46C8-8F1E-9A86EDAE9D86}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{477D705E-576B-46C8-8F1E-9A86EDAE9D86}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{477D705E-576B-46C8-8F1E-9A86EDAE9D86}.Release|Any CPU.Build.0 = Release|Any CPU
 		{33581FDE-ABAF-4C27-A40A-1A2743309399}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{33581FDE-ABAF-4C27-A40A-1A2743309399}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{33581FDE-ABAF-4C27-A40A-1A2743309399}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -230,6 +224,18 @@ Global
 		{34211B07-381E-4489-AA7F-7697D106B7DC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{34211B07-381E-4489-AA7F-7697D106B7DC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{34211B07-381E-4489-AA7F-7697D106B7DC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BC56D244-6C76-4E40-8351-B74B636B17A6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BC56D244-6C76-4E40-8351-B74B636B17A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BC56D244-6C76-4E40-8351-B74B636B17A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BC56D244-6C76-4E40-8351-B74B636B17A6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D96279FD-0020-427B-85C5-DC1C2336AA25}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D96279FD-0020-427B-85C5-DC1C2336AA25}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D96279FD-0020-427B-85C5-DC1C2336AA25}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D96279FD-0020-427B-85C5-DC1C2336AA25}.Release|Any CPU.Build.0 = Release|Any CPU
+		{71F61C77-E04F-464D-A0B7-E109CE7E30D7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{71F61C77-E04F-464D-A0B7-E109CE7E30D7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{71F61C77-E04F-464D-A0B7-E109CE7E30D7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{71F61C77-E04F-464D-A0B7-E109CE7E30D7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -265,8 +271,6 @@ Global
 		{2F7EEC69-35F4-4D31-B52A-7465D65E9DE0} = {A4E67E09-3156-4D30-B451-F24F706E96C4}
 		{E909EEF5-5EBF-43F3-AB4A-6B2FB56EF89B} = {1B0865AF-369E-493C-AA6F-C56D3E520A22}
 		{B43160C8-C9E3-43F6-83E9-A9A45B32F022} = {B5E677CF-2920-4B0A-A056-E73F6B2CF2BC}
-		{4C1CF0E7-4CD3-4731-9A56-E033D5216D58} = {EF6194B2-9ACB-49B9-8049-DD6AFAEB0399}
-		{477D705E-576B-46C8-8F1E-9A86EDAE9D86} = {844CEACD-4C85-4B15-9E2B-892B01FDA4BB}
 		{33581FDE-ABAF-4C27-A40A-1A2743309399} = {2B7F59E8-147F-4399-9804-E7EAEF2DCB22}
 		{5E654407-E22F-4696-A33F-C4B372F547BD} = {613F0980-91ED-4064-8E8C-168582EF4AD7}
 		{8C389735-61CE-405C-972F-3790DF1E823E} = {B5E677CF-2920-4B0A-A056-E73F6B2CF2BC}
@@ -274,6 +278,9 @@ Global
 		{5576D9C5-8856-44AD-86AD-FE73A22751EF} = {53E9CEBA-2D11-41FD-AA60-0151A5442CC4}
 		{0EB6D474-A915-4075-B142-FE07321F3AA9} = {53E9CEBA-2D11-41FD-AA60-0151A5442CC4}
 		{34211B07-381E-4489-AA7F-7697D106B7DC} = {B5E677CF-2920-4B0A-A056-E73F6B2CF2BC}
+		{BC56D244-6C76-4E40-8351-B74B636B17A6} = {EF6194B2-9ACB-49B9-8049-DD6AFAEB0399}
+		{D96279FD-0020-427B-85C5-DC1C2336AA25} = {EF6194B2-9ACB-49B9-8049-DD6AFAEB0399}
+		{71F61C77-E04F-464D-A0B7-E109CE7E30D7} = {844CEACD-4C85-4B15-9E2B-892B01FDA4BB}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {94C0DA2C-CCCB-4314-93A2-9809B5DD0583}
