+++ src/SkyApm.Agent.AspNetCore/SkyApmHostingStartup.cs
@@ -18,6 +18,8 @@
 
 using Microsoft.AspNetCore.Hosting;
 using SkyApm.Agent.AspNetCore;
+using SkyApm.Agent.Hosting;
+using SkyApm.AspNetCore.Diagnostics;
 
 [assembly: HostingStartup(typeof(SkyApmHostingStartup))]
 
@@ -27,7 +29,7 @@ namespace SkyApm.Agent.AspNetCore
     {
         public void Configure(IWebHostBuilder builder)
         {
-            builder.ConfigureServices(services => services.AddSkyAPMCore());
+            builder.ConfigureServices(services => services.AddSkyAPM(ext => ext.AddAspNetCoreHosting()));
         }
     }
 }
